SONAR-9004 Fix max authentication session timeout
Max authentication session timeout was limited to 25 days because of integer usage instead of long

+++ server/sonar-server/src/main/java/org/sonar/server/authentication/JwtSerializer.java
@@ -156,14 +156,14 @@ public class JwtSerializer implements Startable {
   static class JwtSession {
 
     private final String userLogin;
-    private final int expirationTimeInSeconds;
+    private final long expirationTimeInSeconds;
     private final Map<String, Object> properties;
 
-    JwtSession(String userLogin, int expirationTimeInSeconds) {
+    JwtSession(String userLogin, long expirationTimeInSeconds) {
       this(userLogin, expirationTimeInSeconds, Collections.emptyMap());
     }
 
-    JwtSession(String userLogin, int expirationTimeInSeconds, Map<String, Object> properties) {
+    JwtSession(String userLogin, long expirationTimeInSeconds, Map<String, Object> properties) {
       this.userLogin = requireNonNull(userLogin, "User login cannot be null");
       this.expirationTimeInSeconds = expirationTimeInSeconds;
       this.properties = properties;
@@ -173,7 +173,7 @@ public class JwtSerializer implements Startable {
       return userLogin;
     }
 
-    int getExpirationTimeInSeconds() {
+    long getExpirationTimeInSeconds() {
       return expirationTimeInSeconds;
     }
